feat: adds support for settling funds
