+++ components/Proposals/ProposalDetailCard.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useMemo, useState } from 'react';
 import {
-  ActionIcon,
   Button,
   Fieldset,
   Grid,
@@ -10,32 +9,21 @@ import {
   Progress,
   SegmentedControl,
   Stack,
-  Table,
   Text,
   TextInput,
-  useMantineTheme,
 } from '@mantine/core';
 import Link from 'next/link';
-import { useWallet } from '@solana/wallet-adapter-react';
-import { IconExternalLink, IconRefresh, IconTrash, Icon3dRotate } from '@tabler/icons-react';
+import { IconExternalLink } from '@tabler/icons-react';
 import numeral from 'numeral';
-import { BN } from '@coral-xyz/anchor';
 import { useProposal } from '@/hooks/useProposal';
 import { useTokens } from '@/hooks/useTokens';
 import { useTokenAmount } from '@/hooks/useTokenAmount';
-import { TWAPOracle, OpenOrdersAccountWithKey, LeafNode } from '@/lib/types';
+import { TWAPOracle, LeafNode } from '@/lib/types';
 import { NUMERAL_FORMAT } from '@/lib/constants';
-import { useOpenbookTwap } from '@/hooks/useOpenbookTwap';
-import { useTransactionSender } from '@/hooks/useTransactionSender';
-import { useExplorerConfiguration } from '@/hooks/useExplorerConfiguration';
+import { ProposalOrdersCard } from './ProposalOrdersCard';
 
 export function ProposalDetailCard({ proposalNumber }: { proposalNumber: number }) {
-  const theme = useMantineTheme();
-  const { cancelOrderTransactions, settleFundsTransactions } = useOpenbookTwap();
-  const { generateExplorerLink } = useExplorerConfiguration();
-  const sender = useTransactionSender();
-  const wallet = useWallet();
-  const { proposal, markets, orders, mintTokens, placeOrder, loading, fetchOrders } = useProposal({
+  const { proposal, markets, orders, mintTokens, placeOrder, loading } = useProposal({
     fromNumber: proposalNumber,
   });
   const [mintBaseAmount, setMintBaseAmount] = useState<number>();
@@ -60,8 +48,6 @@ export function ProposalDetailCard({ proposalNumber }: { proposalNumber: number 
   const [passPrice, setPassPrice] = useState<number>(0);
   const [failPrice, setFailPrice] = useState<number>(0);
   const [orderType, setOrderType] = useState<string>('Limit');
-  const [isCanceling, setIsCanceling] = useState<boolean>(false);
-  const [isSettling, setIsSettling] = useState<boolean>(false);
 
   const orderbook = useMemo(() => {
     if (!markets) return;
@@ -86,51 +72,6 @@ export function ProposalDetailCard({ proposalNumber }: { proposalNumber: number 
       fail: { asks: getSide(markets.failAsks), bids: getSide(markets.failBids, true) },
     };
   }, [markets]);
-
-  const handleCancel = useCallback(
-    async (order: OpenOrdersAccountWithKey) => {
-      if (!proposal || !markets) return;
-
-      const txs = await cancelOrderTransactions(
-        new BN(order.account.accountNum),
-        proposal.account.openbookPassMarket.equals(order.account.market)
-          ? { publicKey: proposal.account.openbookPassMarket, account: markets.pass }
-          : { publicKey: proposal.account.openbookFailMarket, account: markets.fail },
-      );
-
-      if (!wallet.publicKey || !txs) return;
-
-      try {
-        setIsCanceling(true);
-        await sender.send(txs);
-        setTimeout(() => fetchOrders(), 3000);
-      } finally {
-        setIsCanceling(false);
-      }
-    },
-    [proposal, cancelOrderTransactions, fetchOrders, sender],
-  );
-
-  const handleSettleFunds = useCallback(
-    async (order: OpenOrdersAccountWithKey) => {
-      if (!proposal || !markets) return;
-      const txs = await settleFundsTransactions(
-        new BN(order.account.accountNum),
-        proposal.account.openbookPassMarket.equals(order.account.market)
-          ? { publicKey: proposal.account.openbookPassMarket, account: markets.pass }
-          : { publicKey: proposal.account.openbookFailMarket, account: markets.fail },
-      );
-      if (!wallet.publicKey || !txs) return;
-      try {
-        setIsSettling(true);
-        await sender.send(txs);
-        setTimeout(() => fetchOrders(), 3000);
-      } finally {
-        setIsSettling(false);
-      }
-    },
-    [proposal, settleFundsTransactions, fetchOrders, sender],
-  );
 
   const handleMint = useCallback(
     async (fromBase?: boolean) => {
@@ -389,8 +330,8 @@ export function ProposalDetailCard({ proposalNumber }: { proposalNumber: number 
                 Pass market orderbook
               </Text>
               <Group gap="0">
-                {orderbook.pass.asks?.parsed.map((ask) => (
-                  <Grid w="100%" gutter={0} mih="md">
+                {orderbook.pass.asks?.parsed.map((ask, index) => (
+                  <Grid key={index} w="100%" gutter={0} mih="md">
                     <Grid.Col span={3} />
                     <Grid.Col span={1.5} h="sm" p="0">
                       <Text size="0.6rem">{numeral(ask.price).format(NUMERAL_FORMAT)}</Text>
@@ -409,8 +350,8 @@ export function ProposalDetailCard({ proposalNumber }: { proposalNumber: number 
                     </Grid.Col>
                   </Grid>
                 ))}
-                {orderbook.pass.bids?.parsed.map((bid) => (
-                  <Grid w="100%" gutter={0} mih="md">
+                {orderbook.pass.bids?.parsed.map((bid, index) => (
+                  <Grid key={index} w="100%" gutter={0} mih="md">
                     <Grid.Col span={3}>
                       <Progress
                         key={bid.price + bid.size}
@@ -436,8 +377,8 @@ export function ProposalDetailCard({ proposalNumber }: { proposalNumber: number 
                 Fail market orderbook
               </Text>
               <Group gap="0">
-                {orderbook.fail.asks?.parsed.map((ask) => (
-                  <Grid w="100%" gutter={0} mih="md">
+                {orderbook.fail.asks?.parsed.map((ask, index) => (
+                  <Grid key={index} w="100%" gutter={0} mih="md">
                     <Grid.Col span={3} h="sm" p="0" />
                     <Grid.Col span={1.5} h="sm" p="0">
                       <Text size="0.6rem">{numeral(ask.price).format(NUMERAL_FORMAT)}</Text>
@@ -456,8 +397,8 @@ export function ProposalDetailCard({ proposalNumber }: { proposalNumber: number 
                     </Grid.Col>
                   </Grid>
                 ))}
-                {orderbook.fail.bids?.parsed.map((bid) => (
-                  <Grid w="100%" gutter={0} mih="md">
+                {orderbook.fail.bids?.parsed.map((bid, index) => (
+                  <Grid key={index} w="100%" gutter={0} mih="md">
                     <Grid.Col span={3}>
                       <Progress
                         key={bid.price + bid.size}
@@ -481,121 +422,11 @@ export function ProposalDetailCard({ proposalNumber }: { proposalNumber: number 
           </Group>
         ) : null}
         {proposal && orders ? (
-          <Stack>
-            <Group justify="space-between">
-              <Text fw="bolder" size="xl">
-                Orders
-              </Text>
-              <ActionIcon variant="subtle" onClick={() => fetchOrders()}>
-                <IconRefresh />
-              </ActionIcon>
-            </Group>
-            <Table>
-              <Table.Thead>
-                <Table.Tr>
-                  <Table.Th>Order ID</Table.Th>
-                  <Table.Th>Market</Table.Th>
-                  <Table.Th>Side</Table.Th>
-                  <Table.Th>Quantity</Table.Th>
-                  <Table.Th>Price</Table.Th>
-                  <Table.Th>Amount</Table.Th>
-                  <Table.Th>Actions</Table.Th>
-                </Table.Tr>
-              </Table.Thead>
-              <Table.Tbody>
-                {orders.map((order) => {
-                  const pass = order.account.market.equals(proposal.account.openbookPassMarket);
-                  const bids = order.account.position.bidsBaseLots.gt(
-                    order.account.position.asksBaseLots,
-                  );
-                  return (
-                    (
-                      (order.account.openOrders[0].isFree === 0)
-                    ) ? (
-                     <Table.Tr key={order.publicKey.toString()}>
-                      <Table.Td>
-                        <a href={generateExplorerLink(order.publicKey.toString(), 'account')} target="_blank" rel="noreferrer">
-                          {order.account.accountNum}
-                        </a>
-                      </Table.Td>
-                      <Table.Td c={pass ? theme.colors.green[9] : theme.colors.red[9]}>
-                        {pass ? 'PASS' : 'FAIL'}
-                      </Table.Td>
-                      <Table.Td c={bids ? theme.colors.green[9] : theme.colors.red[9]}>
-                        {bids ? 'BID' : 'ASK'}
-                      </Table.Td>
-                      <Table.Td>
-                        {numeral(
-                          bids
-                            ? order.account.position.bidsBaseLots.toString()
-                            : order.account.position.asksBaseLots.toString(),
-                        ).format(NUMERAL_FORMAT)}
-                      </Table.Td>
-                      <Table.Td>
-                        ${
-                          (parseFloat(order.account.openOrders[0].lockedPrice.toNumber()) / 10000)
-                        }
-                      </Table.Td>
-                      <Table.Td>
-                        ${ bids ?
-                          (
-                            (order.account.position.bidsBaseLots.toNumber()
-                            * order.account.openOrders[0].lockedPrice.toNumber()) / 10000
-                          )
-                          :
-                          (
-                            (order.account.position.asksBaseLots.toNumber()
-                            * order.account.openOrders[0].lockedPrice.toNumber()) / 10000
-                          )
-                        }
-                      </Table.Td>
-                      <Table.Td>
-                        <ActionIcon
-                          variant="subtle"
-                          loading={isCanceling}
-                          onClick={() => handleCancel(order)}
-                        >
-                          <IconTrash />
-                        </ActionIcon>
-                      </Table.Td>
-                     </Table.Tr>)
-                    : (
-                      <Table.Tr key={order.publicKey.toString()}>
-                       <Table.Td>
-                         <a href={generateExplorerLink(order.publicKey.toString(), 'account')} target="_blank" rel="noreferrer">
-                           {order.account.accountNum}
-                         </a>
-                       </Table.Td>
-                       <Table.Td c={pass ? theme.colors.green[9] : theme.colors.red[9]}>
-                         {pass ? 'PASS' : 'FAIL'}
-                       </Table.Td>
-                       <Table.Td c={bids ? theme.colors.green[9] : theme.colors.red[9]}>
-                         {bids ? 'BID' : 'ASK'}
-                       </Table.Td>
-                       <Table.Td>
-                        UNKNOWN
-                       </Table.Td>
-                       <Table.Td>
-                        UNKNOWN
-                       </Table.Td>
-                       <Table.Td>
-                        UNKNOWN
-                       </Table.Td>
-                       <Table.Td>
-                         <ActionIcon
-                           variant="subtle"
-                           loading={isSettling}
-                           onClick={() => handleSettleFunds(order)}
-                         >
-                           <Icon3dRotate />
-                         </ActionIcon>
-                       </Table.Td>
-                      </Table.Tr>)
-                  );
-                })}
-              </Table.Tbody>
-            </Table>
-          </Stack>
+          <ProposalOrdersCard
+            markets={markets}
+            proposal={proposal}
+            orders={orders}
+          />
         ) : null}
       </Stack>
     </Stack>
